move useLocalStorage to /hooks/local-storage
